Free node in error branch
In case the regex cannot be compiled free the allocated node, since it
is not attached to the list.

Also use unsigned long formatter for the pcre2 error offset, since the
logging interface does not support %zu, and print the invalid regular
expression itself.

Also use logging interface instead of stdout.